Break up mn_checks
