+++ src/masternodes/accounts.h
@@ -53,6 +53,88 @@ struct CFuturesUserValue {
     }
 };
 
+struct CAccountToUtxosMessage {
+    CScript from;
+    CBalances balances;
+    uint32_t mintingOutputsStart;
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream &s, Operation ser_action) {
+        READWRITE(from);
+        READWRITE(balances);
+        READWRITE(VARINT(mintingOutputsStart));
+    }
+};
+
+struct CAccountToAccountMessage {
+    CScript from;
+    CAccounts to;  // to -> balances
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream &s, Operation ser_action) {
+        READWRITE(from);
+        READWRITE(to);
+    }
+};
+
+struct CAnyAccountsToAccountsMessage {
+    CAccounts from;  // from -> balances
+    CAccounts to;    // to -> balances
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream &s, Operation ser_action) {
+        READWRITE(from);
+        READWRITE(to);
+    }
+};
+
+struct CUtxosToAccountMessage {
+    CAccounts to;  // to -> balances
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream &s, Operation ser_action) {
+        READWRITE(to);
+    }
+};
+
+struct CSmartContractMessage {
+    std::string name;
+    CAccounts accounts;
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream &s, Operation ser_action) {
+        READWRITE(name);
+        READWRITE(accounts);
+    }
+};
+
+struct CFutureSwapMessage {
+    CScript owner;
+    CTokenAmount source{};
+    uint32_t destination{};
+    bool withdraw{};
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream &s, Operation ser_action) {
+        READWRITE(owner);
+        READWRITE(source);
+        READWRITE(destination);
+        READWRITE(withdraw);
+    }
+};
+
 class CAccountsView : public virtual CStorageView {
 public:
     void ForEachAccount(std::function<bool(const CScript &)> callback, const CScript &start = {});
