+++ src/masternodes/balances.h
@@ -146,89 +146,7 @@ struct CBalances {
     }
 };
 
-struct CAccountToUtxosMessage {
-    CScript from;
-    CBalances balances;
-    uint32_t mintingOutputsStart;
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(from);
-        READWRITE(balances);
-        READWRITE(VARINT(mintingOutputsStart));
-    }
-};
-
 using CAccounts = std::map<CScript, CBalances>;
-
-struct CAccountToAccountMessage {
-    CScript from;
-    CAccounts to;  // to -> balances
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(from);
-        READWRITE(to);
-    }
-};
-
-struct CAnyAccountsToAccountsMessage {
-    CAccounts from;  // from -> balances
-    CAccounts to;    // to -> balances
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(from);
-        READWRITE(to);
-    }
-};
-
-struct CUtxosToAccountMessage {
-    CAccounts to;  // to -> balances
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(to);
-    }
-};
-
-struct CSmartContractMessage {
-    std::string name;
-    CAccounts accounts;
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(name);
-        READWRITE(accounts);
-    }
-};
-
-struct CFutureSwapMessage {
-    CScript owner;
-    CTokenAmount source{};
-    uint32_t destination{};
-    bool withdraw{};
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(owner);
-        READWRITE(source);
-        READWRITE(destination);
-        READWRITE(withdraw);
-    }
-};
 
 inline CBalances SumAllTransfers(const CAccounts &to) {
     CBalances sum;
