+++ src/masternodes/govvariables/attributes.h
@@ -312,6 +312,8 @@ void TrackLiveBalances(CCustomCSView &mnview, const CBalances &balances, const u
 void TrackDUSDAdd(CCustomCSView &mnview, const CTokenAmount &amount);
 void TrackDUSDSub(CCustomCSView &mnview, const CTokenAmount &amount);
 
+Res StoreGovVars(const CGovernanceHeightMessage &obj, CCustomCSView &view);
+
 enum GovVarsFilter {
     All,
     NoAttributes,
